requirements.txt
albumentations==1.2.1
MedPy==0.4.0
numpy==1.21.5
opencv_python_headless==4.5.4.60
Pillow==9.4.0
PyWavelets==1.3.0
scikit_learn==1.2.1
scipy==1.7.3
SimpleITK==2.2.1
torch==1.8.0+cu111
torchio==0.18.84
torchvision==0.9.0+cu111
visdom==0.1.8.9
